requirements.txt
discord.py
aiohttp
aiosqlite
python-dotenv
flask
